Template: use i32 by default
Most of the tasks use values that fit well within an i32.

+++ template/{{cookiecutter.year}}/src/day25.rs
@@ -1,9 +1,9 @@
-pub fn day25p1(input: &[i64]) -> i64 {
-    input.len() as i64
+pub fn day25p1(input: &[i32]) -> i32 {
+    input.len() as i32
 }
 
-pub fn day25p2(input: &[i64]) -> i64 {
-    input.len() as i64
+pub fn day25p2(input: &[i32]) -> i32 {
+    input.len() as i32
 }
 
 #[cfg(test)]
